requirements/base.txt
Django
djangorestframework>=3.8
